requirements.txt
beautifulsoup4==4.8.1
blis==0.4.1
cachetools==3.1.1
catalogue==0.0.8
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cymem==2.0.3
decorator==4.4.1
en-core-web-sm==2.2.5
ffmpeg-python==0.2.0
Flask==1.1.1
Flask-Cors==3.0.8
future==0.18.2
google-api-core==1.14.3
google-auth==1.7.0
google-cloud-core==1.0.3
google-cloud-speech==1.2.0
google-cloud-storage==1.22.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.1
MarkupSafe==1.1.1
more-itertools==7.2.0
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.4
plac==1.1.3
preshed==3.0.2
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pydub==0.23.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.5
spacy==2.2.3
srsly==0.2.0
thinc==7.3.1
tqdm==4.39.0
urllib3==1.25.6
validators==0.14.1
wasabi==0.4.0
Werkzeug==0.16.0
wikipedia==1.4.0
youtube-dl==2019.11.28
zipp==0.6.0